requirements.txt
streamlit
openai
azure-storage-blob
azure-search-documents
azure-core
PyPDF2
python-docx
requests
azure-ai-vision-imageanalysis
Pillow
azure-cognitiveservices-vision-computervision